Praktikum 1_Jobsheet 7_PWL

+++ database/migrations/2026_03_19_092848_add_timestamps_to_categories_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasColumn('categories', 'created_at') || Schema::hasColumn('categories', 'updated_at')) {
+            return;
+        }
+
         Schema::table('categories', function (Blueprint $table) {
             $table->timestamps();
         });
@@ -21,6 +25,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! Schema::hasColumn('categories', 'created_at') && ! Schema::hasColumn('categories', 'updated_at')) {
+            return;
+        }
+
         Schema::table('categories', function (Blueprint $table) {
             $table->dropColumn(['created_at', 'updated_at']);
         });
